refactor(xml): split up dump module
