+++ src/xml_db/dump/mod.rs
@@ -1,3 +1,7 @@
+mod entry;
+mod group;
+mod meta;
+
 use base64::{engine::general_purpose as base64_engine, Engine as _};
 use xml::{
     writer::{EventWriter, XmlEvent as WriterEvent},
@@ -6,7 +10,7 @@ use xml::{
 
 use crate::{
     crypt::ciphers::Cipher,
-    db::{Database, Entry, Group, Meta, Node, Times, Value},
+    db::{Database, Times},
     xml_db::get_epoch_baseline,
 };
 
@@ -113,34 +117,6 @@ impl<S: AsRef<str>, D: DumpXml> DumpXml for SimpleTag<S, D> {
     }
 }
 
-impl DumpXml for Value {
-    fn dump_xml<E: std::io::Write>(
-        &self,
-        writer: &mut EventWriter<E>,
-        inner_cipher: &mut dyn Cipher,
-    ) -> Result<(), xml::writer::Error> {
-        match self {
-            Value::Bytes(b) => SimpleTag("Value", std::str::from_utf8(b).expect("utf-8"))
-                .dump_xml(writer, inner_cipher),
-            Value::Unprotected(s) => SimpleTag("Value", s).dump_xml(writer, inner_cipher),
-            Value::Protected(p) => {
-                writer.write(WriterEvent::start_element("Value").attr("Protected", "True"))?;
-
-                let encrypted_value = inner_cipher
-                    .encrypt(p.unsecure())
-                    .expect("Encrypt with inner cipher");
-
-                let protected_value = base64_engine::STANDARD.encode(&encrypted_value);
-
-                writer.write(WriterEvent::characters(&protected_value))?;
-
-                writer.write(WriterEvent::end_element())?;
-                Ok(())
-            }
-        }
-    }
-}
-
 impl DumpXml for Database {
     fn dump_xml<E: std::io::Write>(
         &self,
@@ -158,121 +134,6 @@ impl DumpXml for Database {
         writer.write(WriterEvent::end_element())?; // Root
 
         writer.write(WriterEvent::end_element())?; // KeePassFile
-
-        Ok(())
-    }
-}
-
-impl DumpXml for Meta {
-    fn dump_xml<E: std::io::Write>(
-        &self,
-        writer: &mut EventWriter<E>,
-        inner_cipher: &mut dyn Cipher,
-    ) -> Result<(), xml::writer::Error> {
-        writer.write(WriterEvent::start_element("Meta"))?;
-
-        SimpleTag("Generator", "keepass-rs").dump_xml(writer, inner_cipher)?;
-
-        // TODO DatabaseName
-        // TODO DatabaseNameChanged
-        // TODO DatabaseDescription
-        // TODO DatabaseDescriptionChanged
-        // TODO DefaultUserName
-        // TODO DefaultUserNameChanged
-        // TODO DeletedObjects
-        // TODO MaintenanceHistoryDays
-        // TODO Color
-        // TODO MasterKeyChanged
-        // TODO MasterKeyChangeRec
-        // TODO MasterKeyChangeForce
-        // TODO MemoryProtection
-        // TODO CustomIcons
-        // TODO RecycleBinEnabled
-        // TODO RecycleBinUUID
-        // TODO RecycleBinChanged
-        // TODO EntryTemplatesGroup
-        // TODO EntryTemplatesGroupChanged
-        // TODO LastSelectedGroup
-        // TODO LastTopVisibleGroup
-        // TODO HistoryMaxItems
-        // TODO HistoryMaxSize
-        // TODO SettingsChanged
-        // TODO CustomData
-
-        writer.write(WriterEvent::end_element())?;
-
-        Ok(())
-    }
-}
-
-impl DumpXml for Node {
-    fn dump_xml<E: std::io::Write>(
-        &self,
-        writer: &mut EventWriter<E>,
-        inner_cipher: &mut dyn Cipher,
-    ) -> Result<(), xml::writer::Error> {
-        match self {
-            Node::Group(g) => g.dump_xml(writer, inner_cipher),
-            Node::Entry(e) => e.dump_xml(writer, inner_cipher),
-        }
-    }
-}
-
-impl DumpXml for Group {
-    fn dump_xml<E: std::io::Write>(
-        &self,
-        writer: &mut EventWriter<E>,
-        inner_cipher: &mut dyn Cipher,
-    ) -> Result<(), xml::writer::Error> {
-        writer.write(WriterEvent::start_element("Group"))?;
-
-        // TODO IconId
-        // TODO Notes
-
-        SimpleTag("Name", &self.name).dump_xml(writer, inner_cipher)?;
-        SimpleTag("UUID", &self.uuid).dump_xml(writer, inner_cipher)?;
-
-        for child in &self.children {
-            child.dump_xml(writer, inner_cipher)?;
-        }
-
-        writer.write(WriterEvent::end_element())?; // Group
-
-        Ok(())
-    }
-}
-
-impl DumpXml for Entry {
-    fn dump_xml<E: std::io::Write>(
-        &self,
-        writer: &mut EventWriter<E>,
-        inner_cipher: &mut dyn Cipher,
-    ) -> Result<(), xml::writer::Error> {
-        writer.write(WriterEvent::start_element("Entry"))?;
-
-        // TODO IconId
-        // TODO Times
-        // TODO AutoType
-        // TODO History
-        // TODO ForegroundColor
-        // TODO BackgroundColor
-
-        SimpleTag("UUID", &self.uuid).dump_xml(writer, inner_cipher)?;
-
-        SimpleTag("Tags", &self.tags.join(";")).dump_xml(writer, inner_cipher)?;
-
-        self.times.dump_xml(writer, inner_cipher)?;
-
-        for (field_name, field_value) in &self.fields {
-            writer.write(WriterEvent::start_element("String"))?;
-
-            SimpleTag("Key", field_name).dump_xml(writer, inner_cipher)?;
-            field_value.dump_xml(writer, inner_cipher)?;
-
-            writer.write(WriterEvent::end_element())?; // String
-        }
-
-        writer.write(WriterEvent::end_element())?; // Entry
 
         Ok(())
     }
